% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLuster.R
\name{CLuster}
\alias{CLuster}
\title{A CLuster sampler using R}
\usage{
CLuster(data, m, M)
}
\arguments{
\item{data}{a data matrix or a dist object}

\item{m}{the result from Fdbscan(data,eps,M)}

\item{M}{number of minimum points in the eps region (for core points)}
}
\value{
a list of Classification of data
}
\description{
Further processing of data based on Fdbscan method
}
\examples{
\dontrun{
data<-iris[,1:2]
m<-Fdbscan(data,0.25,3)
z<-CLuster(data,m,3)
z
}
}
