% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fdbscan.R
\name{Fdbscan}
\alias{Fdbscan}
\title{A Fdbscan sampler using R}
\usage{
Fdbscan(data, eps, M)
}
\arguments{
\item{data}{a data matrix or a dist object}

\item{eps}{size of the epsilon neighborhood}

\item{M}{number of minimum points in the eps region (for core points)}
}
\value{
An object of class 'dbscan_fast'
}
\description{
Fdbscan is a method for preliminary clustering of given data based on DBSCAN algorithm
}
\examples{
\dontrun{
data<-iris[,1:2]
m<-Fdbscan(data,0.25,3)
m
}
}
