% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_C}
\alias{rw_C}
\title{A random walk Metropolis sampler for generating the standard Laplace distribution}
\usage{
rw_C(sigma, x0, N)
}
\arguments{
\item{sigma}{The sample variance}

\item{x0}{Initial value}

\item{N}{Cycles}
}
\value{
A set of data about x
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution
}
