\name{SC19031-package}
\alias{SC19031-package}
\alias{SC19031}
\docType{package}
\title{
  Quick operation based on DBSCAN clustering algorithm
}
\description{
  Fast reimplementation of the DBSCAN (Density-based spatial clustering of applications with noise) clustering algorithm using a kd-tree. The implementation is significantly faster and can work with larger data sets then dbscan in fpc.
}

\author{
Jia Shao

Maintainer: Jia Shao <468574507@qq.com>
}

\keyword{ package }

